Aris entia @spaceapps_ecu
Traveling by plane is essentially travel farther and faster with a cost
according to the needs of our trip. Domestic flights represent a high
percentage of aerial passengers.

Our goal is to simulate the flow of domestic passengers and also help
to improve these so that they can identify areas of interest, general
average costs and consequently investment and subsidies for such
flights.